composites store doubles

+++ t3_ent.h
@@ -19,7 +19,7 @@ class t3_ent
 	virtual t3_sixtuple& getSubEnt(int idx) = 0;
 	virtual std::string str() const = 0;
 	virtual std::string getName() const { return name; }
-	virtual void addSubEnt(int x, int y, int z, int yaw, int roll, const std::string& name) = 0;
+	virtual void addSubEnt(double x, double y, double z, double yaw, double roll, const std::string& name) = 0;
 	virtual bool isContainer() { return false; }
 
 	virtual ~t3_ent() {}
