+++ t3_entComposite.h
@@ -18,7 +18,7 @@ class t3_entComposite : public t3_ent
 	virtual int subEntSize() const;
 	virtual t3_sixtuple& getSubEnt(int idx);
 	virtual std::string str() const;
-	virtual void addSubEnt(int x, int y, int z, int yaw, int roll, const std::string& name);
+	virtual void addSubEnt(double x, double y, double z, double yaw, double roll, const std::string& name);
 	virtual void removeSubEnt(int i);
 	virtual bool isContainer() { return true; }
